+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## [Unreleased]
 
+## 16.2.47 (2018-08-07)
+
+### Grid
+
+#### Bug Fixes
+
+- Batch editing tab key press prevention when grid placed inside dialog fixed.
+- Unable to use `headerText` in group caption template is resolved.
+- Query table name is not used by grid `CRUD` operations fixed.
+- Arguments return properly in `beforeExcelExport` and `beforePdfExport` event.
+- Cancel support for `rowDrop` event is provided.
+
 ## 16.2.46 (2018-07-30)
 
 ### Grid